requirements.txt
amqp==5.2.0
anyascii==0.3.2
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.9.3
billiard==4.2.0
bleach==4.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
defusedxml==0.7.1
Django==3.2.25
django-allauth==0.50.0
django-appconf==1.0.6
django-braces==1.14.0
django-compressor==2.4.1
django-debug-toolbar==3.2.4
django-environ==0.7.0
django-extensions==3.2.3
django-filter==21.1
django-foundation-formtags==0.1.0
django-libsass==0.9
django-modelcluster==5.3
django-redis==5.2.0
django-taggit==2.1.0
django-treebeard==4.7.1
django-wagtail-feeds==0.1.0
djangorestframework==3.15.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.10
kombu==5.4.2
l18n==2021.3
libsass==0.21.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
Pillow==9.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.8.6
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
rcssmin==1.0.6
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.6.1
telepath==0.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wagtail==2.16.3
wagtail-localize==1.1
wagtailmedia==0.10.0
Wand==0.6.7
wcwidth==0.2.13
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0

